+++ database/migrations/2020_02_04_163742_create_user_details_table.php
@@ -14,9 +14,9 @@ class CreateUserDetailsTable extends Migration
     public function up()
     {
         Schema::create('user_details', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('user_id')->unsigned()->nullable();
-            $table->bigInteger('user_group_id')->unsigned()->nullable();
+            $table->increments('id');
+            $table->integer('user_id')->unsigned()->nullable();
+            $table->integer('user_group_id')->unsigned()->nullable();
             $table->integer('status')->nullable()->default(1);
             $table->timestamps();
