+++ database/migrations/2020_02_05_024637_create_settings_table.php
@@ -16,7 +16,7 @@ class CreateSettingsTable extends Migration
         Schema::create('settings', function (Blueprint $table) {
             $table->string('key');
             $table->text('value');
-            $table->bigInteger('updated_by')->unsigned()->nullable();
+            $table->integer('updated_by')->unsigned()->nullable();
             $table->timestamps();
             $table->primary('key');
             $table->foreign('updated_by')->references('id')->on('users')->onDelete('cascade');
